requirements.txt
aiogram==2.25.1
yt-dlp==2025.1.15
ffmpeg-python==0.2.0
redis==5.0.8
python-dotenv==1.0.1
